Correct licensing mistake made
The original/default Model01-Firmware release, from which this is based,
is released under the GPL3 license. My release under the unlicense was a
mistake and, I assume, "null and void" from the get-go.

This commit replaces the unlicense with the correct license, copied
directly from the Arduino IDE's folder for the Model 01 firmware.