Reintroduce Wavepool; refactor/bugfix
Made much easier by the inclusion of Wavepool as a core plugin! :D

Any other non-refactor changes are considered bugfixes because they
bring things in line with where I expected them to be. That is, they are
not simple changes of preference, they are a re-allignment with
preference (or outright bug fixes).

Probably could have been separate commits. Oh well.